BUGFIX: Multiple CORS Upload Widgets on the same form will no longer cause duplicate submissions.
I had not done enough testing with multiple separate CORS Upload Widgets on the
same page, and had thus failed to notice that the S3fsCORSUpload.handleUpload()
function was being executed twice for a single Upload button press if there was
a second CORS Upload Widget on the same form. The code now explicitly binds
handleUpload() to each CORS Widget on the page one at a time.